refactor(example): rename routing module to hub
