+++ example-app/app/views/car/hub/car.hub.ts
@@ -3,7 +3,7 @@ import { RouterModule } from '@angular/router';
 import { createFeature } from 'lib';
 import { CAR_NOTES_KEY, CarNotes } from './car.notes';
 import { carRoutes } from './car.routes';
-import { detailsSlice } from '../../details/hub/details-routing.module';
+import { detailsSlice } from '../../details/hub/details.hub';
 
 export const carSlice = createFeature<CarNotes>(carRoutes, {
   detached: { details: detailsSlice },
@@ -14,4 +14,4 @@ export const carSlice = createFeature<CarNotes>(carRoutes, {
   imports: [RouterModule.forChild(carRoutes)],
   exports: [RouterModule]
 })
-export class CarRoutingModule {}
+export class CarHub {}
