+++ example-app/app/views/details/details.module.ts
@@ -2,10 +2,10 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { DetailsComponent } from './components/details.component';
-import { DetailsRoutingModule } from './hub/details-routing.module';
+import { DetailsHub } from './hub/details.hub';
 
 @NgModule({
   declarations: [DetailsComponent],
-  imports: [CommonModule, DetailsRoutingModule]
+  imports: [CommonModule, DetailsHub]
 })
 export class DetailsModule {}
