+++ example-app/app/views/home/home.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HomeComponent } from './components/home.component';
-import { HomeRoutingModule } from './hub/home-routing.module';
+import { HomeHub } from './hub/home.hub';
 
 @NgModule({
-  imports: [CommonModule, HomeRoutingModule],
+  imports: [CommonModule, HomeHub],
   declarations: [HomeComponent]
 })
 export class HomeModule {}
